pyproject.toml
[tool.poetry]
name = "nonebot_plugin_ddcheck"
version = "0.4.0"
description = "Nonebot2 成分姬插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-ddcheck"
repository = "https://github.com/noneplugin/nonebot-plugin-ddcheck"

[tool.poetry.dependencies]
python = "^3.11"
nonebot2 = "^2.2.0"
nonebot-plugin-alconna = ">=0.38.0,<1.0.0"
nonebot-plugin-htmlrender = "^0.3.0"
nonebot-plugin-apscheduler = ">=0.4.0,<1.0.0"
nonebot-plugin-localstore = ">=0.6.0,<1.0.0"
bilireq = "^0.2.12"
httpx = ">=0.20.0,<1.0.0"
Jinja2 = "^3.0.0"
yt-dlp = "^2024.8.6"
retry = "^0.9.2"
returns = "^0.24.0"
moviepy = "^2.1.2"
drissionpage = "^4.0.5.6"
bilibili-api = {git = "https://github.com/GrahLnn/bilibili-api.git"}

[tool.pyright]
pythonVersion = "3.8"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
